sailface_cloud/requirements.txt
nose
pyjwt[crypto]
